rearrange code structure
