+++ src/py/flwr/client/app.py
@@ -27,8 +27,8 @@ from .grpc_client.connection import insecure_grpc_connection
 from .grpc_client.message_handler import handle
 from .keras_client import KerasClient, KerasClientWrapper
 from .numpy_client import NumPyClient, NumPyClientWrapper
-from flwr.client.sa_client_wrappers import SAClientWrapper, SecAggPlusWrapper
-from flwr.common.light_sec_agg.sa_wrapper import LightSecAggWrapper
+from flwr.client.sa_client_wrappers import LightSecAggWrapper, SecAggPlusWrapper
+from flwr.client.abc_sa_client_wrapper import SAClientWrapper
 
 
 def start_client(
