+++ src/py/flwr/client/sa_client_wrappers/light_sec_agg_sa_wrapper.py
@@ -1,13 +1,13 @@
-from flwr.client.sa_client_wrappers import SAClientWrapper
+from flwr.client.abc_sa_client_wrapper import SAClientWrapper
 from flwr.common.typing import SAServerMessageCarrier, SAClientMessageCarrier
 from flwr.common.timer import Timer
-# from flwr.client import Client
 import flwr.common.light_sec_agg.client_logic as cl
+from flwr.client.client import Client
 
 
 class LightSecAggWrapper(SAClientWrapper):
     """Wrapper which adds LightSecAgg methods."""
-    def __init__(self, c):
+    def __init__(self, c: Client):
         super().__init__(c)
         self.tm = Timer()
 
@@ -22,15 +22,18 @@ class LightSecAggWrapper(SAClientWrapper):
             packet_dict = dict([(str(p[1]), p[2]) for p in res])
             ret_msg = SAClientMessageCarrier(identifier='1', str2scalar=packet_dict)
         elif ins.identifier == '2':
-            packets = [(int(k), self.id, v) for k, v in ins.str2scalar.items()]
+            sec_id = self.get_sec_id()
+            packets = [(int(k), sec_id, v) for k, v in ins.str2scalar.items()]
             res = cl.ask_masked_models(self, packets, ins.fit_ins)
             ret_msg = SAClientMessageCarrier(identifier='2', parameters=res)
         elif ins.identifier == '3':
             active_clients = ins.numpy_ndarray_list[0].tolist()
             res = cl.ask_aggregated_encoded_masks(self, active_clients)
             ret_msg = SAClientMessageCarrier(identifier='3', parameters=res)
+        else:
+            raise Exception("Invalid identifier")
         self.tm.toc('s' + ins.identifier)
-        if self.id == 6:
+        if self.get_sec_id() == 6:
             f = open("log.txt", "a")
             f.write(f"Client without communication stage {ins.identifier}:{self.tm.get('s' + ins.identifier)} \n")
             f.close()
