+++ src/py/flwr/common/light_sec_agg/light_sec_agg_test.py
@@ -4,7 +4,7 @@ import time
 from flwr.common import GRPC_MAX_MESSAGE_LENGTH
 from multiprocessing import Process
 from flwr.server.strategy.light_sec_agg_fedavg import LightSecAggFedAvg
-from flwr.client.sa_client_wrappers.abc_sa_client_wrapper import SAClientWrapper
+from flwr.client.abc_sa_client_wrapper import SAClientWrapper
 from flwr.common.typing import LightSecAggSetupConfigIns, LightSecAggSetupConfigRes, AskEncryptedEncodedMasksIns, \
     EncryptedEncodedMasksPacket, AskMaskedModelsIns, \
     AskAggregatedEncodedMasksIns, SAServerMessageCarrier, SAClientMessageCarrier
