+++ tests/test.py
@@ -89,12 +89,12 @@ if __name__ == "__main__":
                     f.write(
                         f"This is secagg sampling {sample_num} dropping out {dropout_value*5}% with vector size {vector_dimension} try {i} \n")
                     f.close()
-                    sec_agg_test.test_start_simulation(
-                        sample_num=sample_num, share_num=sample_num, threshold=5, vector_dimension=vector_dimension, dropout_value=dropout_value, num_rounds=1)
-                    # light_sec_agg_test.test_start_simulation(
-                    #      sample_num=sample_num,
-                    #      T=int(sample_num * 0.1), U=int(sample_num * 0.7), vector_dimension=vector_dimension, dropout_value=dropout_value
-                    # )
+                    # sec_agg_test.test_start_simulation(
+                    #     sample_num=sample_num, share_num=sample_num, threshold=5, vector_dimension=vector_dimension, dropout_value=dropout_value, num_rounds=1)
+                    light_sec_agg_test.test_start_simulation(
+                         sample_num=sample_num,
+                         T=int(sample_num * 0.1), U=int(sample_num * 0.7), vector_dimension=vector_dimension, dropout_value=dropout_value
+                    )
 
 
 '''# TESTING
